moved top level base routes to their own routes file w/ burnettk
